requirements.txt
GoogleNews
wordcloud
pillow
matplotlib
numpy
pandas
pip